requirements.txt
flask>=3.0.0
numpy>=1.26.0
scipy>=1.12.0
pywebview>=5.0
